[all][xfce4terminal] install less directories
Stow finds out the highest directory that does not exist and creates it
as a symlink.  Problems can crop up in the future when other files that
we don't care about are created in this directory, because in reality it
resides in our git repository.

Reduce the directory hierarchy to be installed to a sane level and
instruct stow to change to the correct directory (install.sh).